refactor: rename InputConsoleView, ResultConsoleView and moved package to impl
